Add a nickserv-based admin verification
This works by keeping track of all admins leaving/joining.

If they join, WHOIS them and add their name to an internal list if they
are identified (numeric 307). If they leave, remove them from that list
because what happens between them leaving and joining again is unknown.

This will probably not work if the bot is in more than one channel.